+++ app/(dashboard)/dashboard/_components/ProtocolsTable/ProtocolsTable.tsx
@@ -1,6 +1,6 @@
 import { DataTable } from '~/components/DataTable/DataTable';
 import { ProtocolColumns } from './Columns';
-import { trpc } from '~/app/_trpc/server';
+import { trpc } from '~/trpc/server';
 
 export const ProtocolsTable = async () => {
   const protocols = await trpc.protocol.get.all.query(undefined, {
