+++ app/(dashboard)/dashboard/_components/UserMenu.tsx
@@ -2,7 +2,7 @@
 
 import { Loader2 } from 'lucide-react';
 import { useRouter } from 'next/navigation';
-import { trpc } from '~/app/_trpc/client';
+import { trpc } from '~/trpc/client';
 import { Button } from '~/components/ui/Button';
 import { useSession } from '~/providers/SessionPrivider';
