+++ app/(dashboard)/dashboard/participants/_components/DeleteAllParticipantsButton.tsx
@@ -2,7 +2,7 @@
 
 import { Loader2 } from 'lucide-react';
 import { useState } from 'react';
-import { trpc } from '~/app/_trpc/client';
+import { trpc } from '~/trpc/client';
 import { Button } from '~/components/ui/Button';
 import {
   AlertDialog,
