+++ app/(dashboard)/dashboard/participants/_components/ImportCSVModal.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useMemo, useState } from 'react';
-import { trpc } from '~/app/_trpc/client';
+import { trpc } from '~/trpc/client';
 import { Button } from '~/components/ui/Button';
 import {
   Dialog,
