+++ app/(onboard)/expired/page.tsx
@@ -4,7 +4,7 @@ import { env } from '~/env.mjs';
 import { userFormClasses } from '../_shared';
 import { Button } from '~/components/ui/Button';
 import { useRouter } from 'next/navigation';
-import { trpc } from '~/app/_trpc/client';
+import { trpc } from '~/trpc/client';
 import { Loader2 } from 'lucide-react';
 
 export default function Page() {
